update polybar launch script

+++ dot_config/polybar/executable_launch.sh
@@ -1,16 +1,38 @@
 #!/bin/sh
 
-killall -q polybar
+echo "Starting polybars"
 
-case $DESKTOP_SESSION in
-	bspwm)  bar=bspwm   ;;
-	xmonad) bar=xmonad  ;;
-	i3)     bar=i3      ;;
-	*)      bar=""      ;;
+# shutdown polybars
+if pgrep -u "$(id -u)" -x polybar >/dev/null; then
+  echo "killing existing polybars"
+  polybar-msg cmd quit
+  sleep 1
+  killall -q polybar
+
+  # wait until shutdown of existing polybars
+  while pgrep -u "$(id -u)" -x polybar >/dev/null; do
+    sleep 0.5;
+  done
+fi
+
+case "$DESKTOP_SESSION" in
+  bspwm)  bar=bspwm  ;;
+  xmonad) bar=xmonad ;;
+  i3)     bar=i3     ;;
+  *)      bar=""     ;;
 esac
 
 [ -z "$bar" ] && exit 1
 
-for m in $(polybar --list-monitors | cut -d":" -f1); do
-  MONITOR=$m polybar --reload "$bar" &
+# display tray only on primary monitor
+monitor_configs="$(LANG=C polybar --list-monitors | sed '
+s/:.*primary.*$/|right/g
+s/:.*$/|none/g
+')"
+
+for mc in $monitor_configs; do
+  monitor="${mc%|*}"
+  tray_position="${mc#*|}"
+
+  POLYBAR_MONITOR="$monitor" POLYBAR_TRAY_POSITION="$tray_position" polybar "$bar" &
 done
